+++ lib/Models/Cesium.ts
@@ -90,6 +90,7 @@ import TerriaFeature from "./Feature/Feature";
 import GlobeOrMap from "./GlobeOrMap";
 import Terria from "./Terria";
 import UserDrawing from "./UserDrawing";
+import { setViewerMode } from "./ViewerMode";
 
 //import Cesium3DTilesInspector from "terriajs-cesium/Source/Widgets/Cesium3DTilesInspector/Cesium3DTilesInspector";
 
@@ -152,6 +153,7 @@ export default class Cesium extends GlobeOrMap {
     });
   });
 
+  private _terrainMessageViewed: boolean = false;
   constructor(terriaViewer: TerriaViewer, container: string | HTMLElement) {
     super();
 
@@ -234,26 +236,6 @@ export default class Cesium extends GlobeOrMap {
     (<any>this.scene).highDynamicRange = false;
 
     this.scene.imageryLayers.removeAll();
-
-    //catch Cesium terrain provider down and switch to Ellipsoid
-    //     terrainProvider.errorEvent.addEventListener(function(err) {
-    //         console.log('Terrain provider error.  ', err.message);
-    //         if (viewer.scene.terrainProvider instanceof CesiumTerrainProvider) {
-    //             console.log('Switching to EllipsoidTerrainProvider.');
-    //             that.terria.viewerMode = ViewerMode.CesiumEllipsoid;
-    //             if (!defined(that.TerrainMessageViewed)) {
-    //                 that.terria.raiseErrorToUser({
-    //                     title : 'Terrain Server Not Responding',
-    //                     message : '\
-    // The terrain server is not responding at the moment.  You can still use all the features of '+that.terria.appName+' \
-    // but there will be no terrain detail in 3D mode.  We\'re sorry for the inconvenience.  Please try \
-    // again later and the terrain server should be responding as expected.  If the issue persists, please contact \
-    // us via email at '+that.terria.supportEmail+'.'
-    //                 });
-    //                 that.TerrainMessageViewed = true;
-    //             }
-    //         }
-    //     });
 
     this.updateCredits(container);
 
@@ -431,6 +413,54 @@ export default class Cesium extends GlobeOrMap {
       this.cesiumWidget.scene.globe.maximumScreenSpaceError =
         this.terria.baseMaximumScreenSpaceError;
     });
+  }
+
+  /** Add an event listener to a TerrainProvider.
+   * If we get an error when trying to load the terrain, then switch to smooth mode, and notify the user.
+   * Finally, remove the listener, so failed tiles do not trigger the error as these can be common and are not a problem. */
+  private async catchTerrainProviderDown(terrainProvider: TerrainProvider) {
+    // Some network errors are not rejected through readyPromise, so we have to
+    // listen to them using the error event and dispose it later
+    let networkErrorListener: (err: any) => void;
+    const networkErrorPromise = new Promise((_resolve, reject) => {
+      networkErrorListener = reject;
+      terrainProvider.errorEvent.addEventListener(networkErrorListener);
+    });
+
+    const isReady = await Promise.race([
+      networkErrorPromise,
+      terrainProvider.readyPromise
+    ])
+      .then(() => {
+        /** Need to throw an error if incorrect `cesiumTerrainUrl` has been specified.
+        The terrainProvider.readyPromise will still be fulfulled, but the map will not load correctly
+        So we check for terrainProvider.availability */
+
+        if (!terrainProvider.availability) {
+          throw new Error();
+        }
+      })
+      .catch((err) => {
+        console.log("Terrain provider error.  ", err.message);
+        if (this.scene.terrainProvider instanceof CesiumTerrainProvider) {
+          console.log("Switching to EllipsoidTerrainProvider.");
+          setViewerMode("3dsmooth", this.terriaViewer);
+          if (!this._terrainMessageViewed) {
+            this.terria.raiseErrorToUser(err, {
+              title: i18next.t("map.cesium.terrainServerErrorTitle"),
+              message: i18next.t("map.cesium.terrainServerErrorMessage", {
+                appName: this.terria.appName,
+                supportEmail: this.terria.supportEmail
+              })
+            });
+
+            this._terrainMessageViewed = true;
+          }
+        }
+      })
+      .finally(() =>
+        terrainProvider.errorEvent.removeEventListener(networkErrorListener)
+      );
   }
 
   private updateCredits(container: string | HTMLElement) {
@@ -945,7 +975,7 @@ export default class Cesium extends GlobeOrMap {
     });
 
     const disposeSplitDirectionChange = autorun(() => {
-      const items = this.terria.mainViewer.items.get();
+      const items = this.terriaViewer.items.get();
       const showSplitter = this.terria.showSplitter;
       items.forEach((item) => {
         if (
@@ -1154,11 +1184,14 @@ export default class Cesium extends GlobeOrMap {
     assetId: number,
     accessToken?: string
   ): CesiumTerrainProvider {
-    return new CesiumTerrainProvider({
+    const terrainProvider = new CesiumTerrainProvider({
       url: IonResource.fromAssetId(assetId, {
         accessToken
       })
     });
+    // Add the event handler to the TerrainProvider
+    this.catchTerrainProviderDown(terrainProvider);
+    return terrainProvider;
   }
 
   /**
@@ -1167,9 +1200,13 @@ export default class Cesium extends GlobeOrMap {
    * Used for spying in specs
    */
   private createTerrainProviderFromUrl(url: string): CesiumTerrainProvider {
-    return new CesiumTerrainProvider({
+    const terrainProvider = new CesiumTerrainProvider({
       url
     });
+
+    // Add the event handler to the TerrainProvider
+    this.catchTerrainProviderDown(terrainProvider);
+    return terrainProvider;
   }
 
   /**
@@ -1178,7 +1215,10 @@ export default class Cesium extends GlobeOrMap {
    * Used for spying in specs
    */
   private createWorldTerrain(): CesiumTerrainProvider {
-    return createWorldTerrain({});
+    const terrainProvider = createWorldTerrain({});
+    // Add the event handler to the TerrainProvider
+    this.catchTerrainProviderDown(terrainProvider);
+    return terrainProvider;
   }
 
   @computed